refactor(model): 将日志和租户模型中的 uint64 类型改为 uint 类型
- 修改了 LoginLog、OperationLogProfile、LoginLogProfile 和 TenantProfile 结构体中的 ID、TenantID 和 UserID 字段类型
- 从 uint64 改为 uint，以减少存储空间需求并提高性能

+++ internal/model/log.go
@@ -31,9 +31,9 @@ func (OperationLog) TableName() string {
 
 // LoginLog 登录日志模型
 type LoginLog struct {
-	ID        uint64    `json:"id" gorm:"primarykey"`
-	TenantID  uint64    `json:"tenant_id" gorm:"not null;default:0;index:idx_tenant_id"`
-	UserID    *uint64   `json:"user_id" gorm:"index:idx_user_id"`
+	ID        uint      `json:"id" gorm:"primarykey"`
+	TenantID  uint      `json:"tenant_id" gorm:"not null;default:0;index:idx_tenant_id"`
+	UserID    *uint     `json:"user_id" gorm:"index:idx_user_id"`
 	Username  string    `json:"username" gorm:"not null;size:50;index:idx_username" validate:"required,max=50"`
 	IP        string    `json:"ip" gorm:"not null;size:45" validate:"required,max=45"`
 	UserAgent string    `json:"user_agent" gorm:"size:500" validate:"max=500"`
@@ -55,9 +55,9 @@ func (LoginLog) TableName() string {
 
 // OperationLogProfile 操作日志资料（简化版本）
 type OperationLogProfile struct {
-	ID           uint64    `json:"id"`
-	TenantID     uint64    `json:"tenant_id"`
-	UserID       uint64    `json:"user_id"`
+	ID           uint      `json:"id"`
+	TenantID     uint      `json:"tenant_id"`
+	UserID       uint      `json:"user_id"`
 	Username     string    `json:"username"`
 	Operation    string    `json:"operation"`
 	Method       string    `json:"method"`
@@ -95,9 +95,9 @@ func (ol *OperationLog) ToProfile() OperationLogProfile {
 
 // LoginLogProfile 登录日志资料（简化版本）
 type LoginLogProfile struct {
-	ID        uint64    `json:"id"`
-	TenantID  uint64    `json:"tenant_id"`
-	UserID    *uint64   `json:"user_id"`
+	ID        uint      `json:"id"`
+	TenantID  uint      `json:"tenant_id"`
+	UserID    *uint     `json:"user_id"`
 	Username  string    `json:"username"`
 	IP        string    `json:"ip"`
 	UserAgent string    `json:"user_agent"`
